+++ .idea/workspace.xml
@@ -2,16 +2,7 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="5f634032-7df3-46e7-b21c-7f7e14cd7923" name="Changes" comment="">
-      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/combined-approach/recall_for_df.py" beforeDir="false" afterPath="$PROJECT_DIR$/combined-approach/code/recall_for_df.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/combined-approach/recall_for_list.py" beforeDir="false" afterPath="$PROJECT_DIR$/combined-approach/code/recall_for_list.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/metric_learning/metric_learning_train.py" beforeDir="false" afterPath="$PROJECT_DIR$/metric_learning/code/metric_learning_train.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/metric_learning/metric_prep.py" beforeDir="false" afterPath="$PROJECT_DIR$/metric_learning/code/metric_prep.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/metric_learning/metric_scoring.py" beforeDir="false" afterPath="$PROJECT_DIR$/metric_learning/code/metric_scoring.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/network-approach/graph.py" beforeDir="false" afterPath="$PROJECT_DIR$/network-approach/code/graph.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/network-approach/network_based_LR.py" beforeDir="false" afterPath="$PROJECT_DIR$/network-approach/code/network_based_LR.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/text-approach/text_based_ensemble_ranker_score.csv" beforeDir="false" afterPath="$PROJECT_DIR$/text-approach/output_data/text_based_ensemble_ranker_score.csv" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/text-approach/text_based_simple_ranker_score.csv" beforeDir="false" afterPath="$PROJECT_DIR$/text-approach/output_data/text_based_simple_ranker_score.csv" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/combined-approach/data/combined_approach.py" beforeDir="false" afterPath="$PROJECT_DIR$/combined-approach/data/combined_approach.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -40,6 +31,7 @@
 }</component>
   <component name="RecentsManager">
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="$PROJECT_DIR$/combined-approach/data" />
       <recent name="$PROJECT_DIR$/combined-approach/code" />
       <recent name="$PROJECT_DIR$/metric_learning/code" />
       <recent name="$PROJECT_DIR$/network-approach/code" />
@@ -59,7 +51,7 @@
       <option name="IS_MODULE_SDK" value="true" />
       <option name="ADD_CONTENT_ROOTS" value="true" />
       <option name="ADD_SOURCE_ROOTS" value="true" />
-      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/combined-approach/data/combined_approach.py" />
+      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/combined-approach/code/combined_approach.py" />
       <option name="PARAMETERS" value="" />
       <option name="SHOW_COMMAND_LINE" value="false" />
       <option name="EMULATE_TERMINAL" value="false" />
